src/models/ML-prediction/requirements.txt
pandas
psycopg2-binary
joblib
scikit-learn